backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
